requirements.txt
alabaster==0.7.13
awscli==1.16.201
Babel==2.12.1
blis==0.7.9
botocore==1.12.191
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.3.9
confection==0.1.0
coverage==7.2.7
cymem==2.0.7
docutils==0.20.1
flake8==6.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
Jinja2==3.1.2
jmespath==0.10.0
langcodes==3.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
murmurhash==1.0.9
numpy==1.24.4
packaging==23.1
pathy==0.10.2
preshed==3.0.8
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.10
pyflakes==3.0.1
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.1
requests==2.31.0
rsa==3.4.2
s3transfer==0.2.1
six==1.16.0
smart-open==6.3.0
snowballstemmer==2.2.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
-e git+ssh://git@git.corp.adobe.com/experience-platform/knowledge-graph-constructor.git@869c17ff706377081cd29fb4d2004fb7887e0b70#egg=src
srsly==2.4.6
thinc==8.1.10
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
urllib3==1.25.11
wasabi==1.1.2
zipp==3.15.0
